Adding new tables according to the new questionnaire; fixed bug in the data insertion in the db
BUG-fix: when inserting data in the surveys table, the records need to get the survey id from the survey table. The insertion process was not waiting for the second to be inserted, so sometimes the data was skipped.

There is still a less urgent fix to do: to retrieve the survey id, a search based on time is done against the db. If we have to users who started the survey at the same time, we may mix things (same problem for http info). The information needed to avoid this problem is in place, we need to make use of it